Base Docker image for Shibboleth IdP
To make this whole thing more reusable, this commit introduces so called
"base docker image". That's the base image from which all the hosts are
to be created.

This image will just contain Java, Jetty and Shibboleth IdP installed,
BUT NOT CONFIGURED. All the actual configuration is going to be done on
a per-host basis.